+++ firebase-crashlytics/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Unreleased
+
+
+# 19.4.4
 * [fixed] Fixed more strict mode violations
+
+
+## Kotlin
+The Kotlin extensions library transitively includes the updated
+`firebase-crashlytics` library. The Kotlin extensions library has no additional
+updates.
 
 # 19.4.3
 * [fixed] Fixed UnbufferedIoViolation strict mode violation [#6822]